--- a/web/models/all.ex
+++ b/web/models/all.ex
@@ -7,7 +7,8 @@ defmodule Isucon5q.User do
     field :nick_name
     field :email
 
-    has_one :salt, Salt, foreign_key: :user_id
+    has_one :salt, Isucon5q.Salt, foreign_key: :user_id
+    has_one :profile, Isucon5q.Profile, foreign_key: :user_id
   end
 
   def get_by(user_id: id) do
@@ -37,6 +38,9 @@ defmodule Isucon5q.Relation do
     field :another, :integer
 
     timestamps inserted_at: :created_at, updated_at: nil
+
+    has_one :user_one, Isucon5q.User, references: :one, foreign_key: :id
+    has_one :user_another, Isucon5q.User, references: :another, foreign_key: :id
   end
 
   def changeset(model, params \\ :empty) do
@@ -101,7 +105,8 @@ defmodule Isucon5q.Profile do
     field :pref
 
     timestamps(inserted_at: nil, updated_at: :updated_at)
-    # field :updated_at, Ecto.DateTime
+
+    belongs_to :user, Isucon5q.User, references: :id, define_field: false
   end
 
   # ref: http://hexdocs.pm/ecto/Ecto.Changeset.html#cast/4
@@ -123,6 +128,9 @@ defmodule Isucon5q.Footprint do
     field :updated_at, Ecto.DateTime, virtual: true
 
     timestamps inserted_at: :created_at, updated_at: nil
+
+    has_one :user, Isucon5q.User, references: :user_id, foreign_key: :id
+    has_one :owner, Isucon5q.User, references: :owner_id, foreign_key: :id
   end
 
   # TODO:
@@ -172,7 +180,8 @@ defmodule Isucon5q.Entry do
 
     timestamps(inserted_at: :created_at, updated_at: nil)
 
-    # has_many :comments, Isucon5q.Comment, foreign_key: :entry_id, references: :id
+    belongs_to :user, Isucon5q.User, foreign_key: :user_id, define_field: false
+    has_many :comments, Isucon5q.Comment
   end
 
   def changeset(model, params \\ :empty) do
@@ -225,7 +234,9 @@ defmodule Isucon5q.Comment do
 
     timestamps inserted_at: :created_at, updated_at: nil
 
-    # belongs_to :entry, Isucon5q.Entry, define_field: false #foreign_key: :entry_id
+    belongs_to :entry, Isucon5q.Entry, define_field: false #foreign_key: :entry_id
+    has_one :user, Isucon5q.User, foreign_key: :id, references: :user_id
+    has_one :entry_user, through: [:entry, :user]
   end
 
   def changeset(model, params \\ :empty) do