--- a/web/templates/page/friends.html.eex
+++ b/web/templates/page/friends.html.eex
@@ -2,9 +2,8 @@
 
 <div class="row panel panel-primary" id="friends">
   <dl>
-    <%= for friend_id <- @friends do %>
-      <% friend = user(friend_id) %>
-      <dt class="friend-date">{{ friend.created_at }}</dt><dd class="friend-friend"><a href="/profile/<%= friend.account_name %>"><%= friend.nick_name %></a></dd>
+    <%= for user <- @friends do %>
+      <dt class="friend-date">{{ friend.created_at }}</dt><dd class="friend-friend"><a href="/profile/<%= user.account_name %>"><%= user.nick_name %></a></dd>
     <% end %>
   </dl>
 </div>