--- a/web/templates/page/index.html.eex
+++ b/web/templates/page/index.html.eex
@@ -31,8 +31,7 @@
     <div id="footprints">
       <ul class="list-group">
         <%= for footprint <- @footprints do %>
-          <% foot_user = user(footprint.user_id) %>
-          <li class="list-group-item footprints-footprint"><%= footprint.created_at %>: <a href="/profile/<%= foot_user.account_name %>"><%= foot_user.nick_name %>さん</a></li>
+          <li class="list-group-item footprints-footprint"><%= footprint.created_at %>: <a href="/profile/<%= footprint.user.account_name %>"><%= footprint.user.nick_name %>さん</a></li>
         <% end %>
       </ul>
     </div>
@@ -46,8 +45,7 @@
       <%= for comment <- @comments do %>
       <div class="comments-comment">
         <ul class="list-group">
-          <% comment_user = user(comment.user_id) %>
-          <li class="list-group-item comment-owner"><a href="/profile/<%= comment_user.account_name %>"><%= comment_user.account_name %>さん</a>:</li>
+          <li class="list-group-item comment-owner"><a href="/profile/<%= comment.user.account_name %>"><%= comment.user.account_name %>さん</a>:</li>
           <li class="list-group-item comment-comment"><%= comment_digest(comment.comment) %></li>
           <li class="list-group-item comment-created-at">投稿時刻:<%= comment.created_at %></li>
         </ul>
@@ -62,8 +60,7 @@
       <%= for entry <- @friend_entries do %>
       <div class="friend-entry">
         <ul class="list-group">
-          <% entry_owner = user(entry.user_id) %>
-          <li class="list-group-item entry-owner"><a href="/diary/entries/<%= entry_owner.account_name %>"><%= entry_owner.nick_name %>さん</a>:</li>
+          <li class="list-group-item entry-owner"><a href="/diary/entries/<%= entry.user.account_name %>"><%= entry.user.nick_name %>さん</a>:</li>
           <li class="list-group-item entry-title"><a href="/diary/entry/<%= entry.id %>"><%= entry_title(entry) %></a></li>
           <li class="list-group-item entry-created-at">投稿時刻:<%= entry.created_at %></li>
         </ul>
@@ -78,10 +75,7 @@
       <%= for comment <- @friend_comments do %>
       <div class="friend-comment">
         <ul class="list-group">
-          <% comment_owner = user(comment.user_id) %>
-          <% comment_entry = entry(comment.entry_id) %>
-          <% entry_owner = user(comment_entry.user_id) %>
-          <li class="list-group-item comment-from-to"><a href="/profile/<%= comment_owner.account_name %>"><%= comment_owner.nick_name %>さん</a>から<a href="/profile/<%= entry_owner.account_name %>"><%= entry_owner.nick_name %>さん</a>へのコメント:</li>
+          <li class="list-group-item comment-from-to"><a href="/profile/<%= comment.user.account_name %>"><%= comment.user.nick_name %>さん</a>から<a href="/profile/<%= comment.entry_user.account_name %>"><%= comment.entry_user.nick_name %>さん</a>へのコメント:</li>
           <li class="list-group-item comment-comment"><%= comment_digest(comment.comment) %></li>
           <li class="list-group-item comment-created-at">投稿時刻:<%= comment.created_at %></li>
         </ul>