4.5x faster.

## SCORE

### before: score=291.6

```
{
  "valid" : true,
  "requests" : 391,
  "elapsed" : 72761,
  "done" : "[{Isucon5InitExecutor},{BootstrapChecker},{Isucon5Load,Isucon5Load,Isucon5Load,Isucon5Load,Isucon5Checker}]",
  "responses" : {
    "success" : 281,
    "redirect" : 106,
    "failure" : 4,
    "error" : 0,
    "exception" : 0
  },
  "violations" : [ ]
}

BUILD SUCCESSFUL

Total time: 1 mins 18.486 secs
```

```
irb(main):001:0> 281 + 106 * 0.1
=> 291.6
```

### after: score=1327.8

```
{
  "valid" : true,
  "requests" : 1714,
  "elapsed" : 66143,
  "done" : "[{Isucon5InitExecutor},{BootstrapChecker},{Isucon5Load,Isucon5Load,Isucon5Load,Isucon5Load,Isucon5Checker}]",
  "responses" : {
    "success" : 1285,
    "redirect" : 428,
    "failure" : 1,
    "error" : 0,
    "exception" : 0
  },
  "violations" : [ ]
}

BUILD SUCCESSFUL

Total time: 1 mins 12.167 secs
```

```
irb(main):002:0> 1285 + 428 * 0.1
=> 1327.8
```

### NOTE

* Latency of index page is still high.
  * 150ms ~ 200ms
* In every request, User.get_by/1 is called at least twice.
  * `Router.authenticate/2` and `PageController.current_user/1`